# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ActionRequestDto": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "cancel",
              "retry"
            ],
            "title": "Action",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "action",
          "requested_by",
          "requested_at"
        ],
        "title": "ActionRequestDto",
        "type": "object"
      },
...
      "CameraPlan": {
        "additionalProperties": false,
        "properties": {
          "focus_x": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Focus X",
            "type": "number"
          },
          "focus_y": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Focus Y",
            "type": "number"
          },
          "mode": {
            "default": "static",
            "enum": [
              "static",
              "push",
              "pan",
              "spotlight"
            ],
            "title": "Mode",
            "type": "string"
          },
          "scale_end": {
            "default": 1.0,
            "maximum": 1.08,
            "minimum": 1.0,
            "title": "Scale End",
            "type": "number"
          },
          "scale_start": {
            "default": 1.0,
            "maximum": 1.08,
            "minimum": 1.0,
            "title": "Scale Start",
            "type": "number"
          }
        },
        "title": "CameraPlan",
        "type": "object"
      },
      "CardStackPayload": {
        "additionalProperties": false,
        "properties": {
          "cards": {
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Cards",
            "type": "array"
          },
          "kind": {
            "const": "card_stack",
            "default": "card_stack",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "cards"
        ],
        "title": "CardStackPayload",
        "type": "object"
      },
...
      "EffectCue": {
        "additionalProperties": false,
        "properties": {
          "end_ms": {
            "exclusiveMinimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "default": "content",
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "start_ms": {
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "start_ms",
          "end_ms"
        ],
        "title": "EffectCue",
        "type": "object"
      },
      "EffectEvent": {
        "additionalProperties": false,
        "properties": {
          "end_ms": {
            "exclusiveMinimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "intensity": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Intensity",
            "type": "number"
          },
          "start_ms": {
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "type": {
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "start_ms",
          "end_ms"
        ],
        "title": "EffectEvent",
        "type": "object"
      },
...
      "EffectPlanV2": {
        "additionalProperties": false,
        "properties": {
          "aspect_ratio": {
            "default": "16:9",
            "enum": [
              "16:9",
              "9:16"
            ],
            "title": "Aspect Ratio",
            "type": "string"
          },
          "background_preset": {
            "default": "tech_blue",
            "enum": [
              "tech_blue",
              "risk_red",
              "warm_gold",
              "paper_grid",
              "regional_teal"
            ],
            "title": "Background Preset",
            "type": "string"
          },
          "camera": {
            "$ref": "#/components/schemas/CameraPlan"
          },
          "cues": {
            "items": {
              "$ref": "#/components/schemas/EffectCue"
            },
            "title": "Cues",
            "type": "array"
          },
          "duration_ms": {
            "exclusiveMinimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "effects": {
            "items": {
              "$ref": "#/components/schemas/EffectEvent"
            },
            "title": "Effects",
            "type": "array"
          },
          "fallback": {
            "$ref": "#/components/schemas/FallbackPlan"
          },
          "legacy_payload_hash": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacy Payload Hash"
          },
          "manual_lock": {
            "default": false,
            "title": "Manual Lock",
            "type": "boolean"
          },
          "migration_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Migration Version"
          },
          "page_id": {
            "minLength": 1,
            "title": "Page Id",
            "type": "string"
          },
          "page_type": {
            "minLength": 1,
            "title": "Page Type",
            "type": "string"
          },
          "presenter_cues": {
            "items": {
              "$ref": "#/components/schemas/PresenterCue"
            },
            "title": "Presenter Cues",
            "type": "array"
          },
          "rhythm_profile": {
            "default": "steady",
            "enum": [
              "steady",
              "standard",
              "compact",
              "enhanced"
            ],
            "title": "Rhythm Profile",
            "type": "string"
          },
          "schema_version": {
            "const": "2.0",
            "default": "2.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source_hashes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Source Hashes",
            "type": "object"
          },
          "template": {
            "default": "SafeSlide",
            "enum": [
              "ProgressiveReveal",
              "ChapterCurtain",
              "StatCounter",
              "ChartNarration",
              "CompareMode",
              "FocusSpotlight",
              "CardStack",
              "GaugeAndRatio",
              "PathBuilder",
              "TagMatrix",
              "RiskAlert",
              "MapHighlight",
              "SafeSlide"
            ],
            "title": "Template",
            "type": "string"
          },
          "template_payload": {
            "discriminator": {
              "mapping": {
                "card_stack": "#/components/schemas/CardStackPayload",
                "chapter_curtain": "#/components/schemas/ChapterCurtainPayload",
                "chart_narration": "#/components/schemas/ChartNarrationPayload",
                "compare_mode": "#/components/schemas/CompareModePayload",
                "focus_spotlight": "#/components/schemas/FocusSpotlightPayload",
                "gauge_and_ratio": "#/components/schemas/GaugeAndRatioPayload",
                "map_highlight": "#/components/schemas/MapHighlightPayload",
                "path_builder": "#/components/schemas/PathBuilderPayload",
                "progressive_reveal": "#/components/schemas/ProgressiveRevealPayload",
                "risk_alert": "#/components/schemas/RiskAlertPayload",
                "safe_slide": "#/components/schemas/SafeSlidePayload",
                "stat_counter": "#/components/schemas/StatCounterPayload",
                "tag_matrix": "#/components/schemas/TagMatrixPayload"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/SafeSlidePayload"
              },
              {
                "$ref": "#/components/schemas/ProgressiveRevealPayload"
              },
              {
                "$ref": "#/components/schemas/ChapterCurtainPayload"
              },
              {
                "$ref": "#/components/schemas/StatCounterPayload"
              },
              {
                "$ref": "#/components/schemas/ChartNarrationPayload"
              },
              {
                "$ref": "#/components/schemas/CompareModePayload"
              },
              {
                "$ref": "#/components/schemas/FocusSpotlightPayload"
              },
              {
                "$ref": "#/components/schemas/CardStackPayload"
              },
              {
                "$ref": "#/components/schemas/GaugeAndRatioPayload"
              },
              {
                "$ref": "#/components/schemas/PathBuilderPayload"
              },
              {
                "$ref": "#/components/schemas/TagMatrixPayload"
              },
              {
                "$ref": "#/components/schemas/RiskAlertPayload"
              },
              {
                "$ref": "#/components/schemas/MapHighlightPayload"
              }
            ],
            "title": "Template Payload"
          },
          "transition": {
            "$ref": "#/components/schemas/TransitionPlan"
          }
        },
        "required": [
          "page_id",
          "page_type",
          "duration_ms"
        ],
        "title": "EffectPlanV2",
        "type": "object"
      },
...
      "Envelope_dict_str__object__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "Envelope[dict[str, object]]",
        "type": "object"
      },
...
      "FallbackPlan": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "template": {
            "default": "SafeSlide",
            "enum": [
              "SafeSlide",
              "FocusSpotlight"
            ],
            "title": "Template",
            "type": "string"
          }
        },
        "title": "FallbackPlan",
        "type": "object"
      },
...
      "GenerateEffectsRequest": {
        "additionalProperties": false,
        "properties": {
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "page_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Ids"
          }
        },
        "title": "GenerateEffectsRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PresenterCue": {
        "additionalProperties": false,
        "properties": {
          "end_ms": {
            "exclusiveMinimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "position": {
            "default": "bottom-right",
            "enum": [
              "top-left",
              "top-right",
              "bottom-left",
              "bottom-right"
            ],
            "title": "Position",
            "type": "string"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "start_ms": {
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          }
        },
        "required": [
          "start_ms",
          "end_ms"
        ],
        "title": "PresenterCue",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "blocking": {
            "title": "Blocking",
            "type": "boolean"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "page_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Id"
          }
        },
        "required": [
          "code",
          "message",
          "action",
          "blocking"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "UnlockEffectRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "UnlockEffectRequest",
        "type": "object"
      },
...
      "UpdateEffectRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          },
          "locked": {
            "default": false,
            "title": "Locked",
            "type": "boolean"
          },
          "plan": {
            "$ref": "#/components/schemas/EffectPlanV2"
          }
        },
        "required": [
          "expected_revision",
          "plan"
        ],
        "title": "UpdateEffectRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PPT Video Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/peripheral/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_peripheral_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Api Peripheral Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/api/peripheral/jobs/{job_id}/actions": {
      "post": {
        "operationId": "request_action_api_peripheral_jobs__job_id__actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Request Action Api Peripheral Jobs  Job Id  Actions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Action"
      }
    },
...
    "/api/projects/{project_id}/effects": {
      "get": {
        "operationId": "workspace_api_projects__project_id__effects_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace"
      }
    },
    "/api/projects/{project_id}/effects/catalog": {
      "get": {
        "operationId": "catalog_api_projects__project_id__effects_catalog_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__object__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Catalog"
      }
    },
    "/api/projects/{project_id}/effects/generate": {
      "post": {
        "operationId": "generate_api_projects__project_id__effects_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GenerateEffectsRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate"
      }
    },
    "/api/projects/{project_id}/effects/pages/{page_id}": {
      "put": {
        "operationId": "update_page_api_projects__project_id__effects_pages__page_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEffectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Page"
      }
    },
    "/api/projects/{project_id}/effects/pages/{page_id}/unlock": {
      "post": {
        "operationId": "unlock_page_api_projects__project_id__effects_pages__page_id__unlock_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockEffectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unlock Page"
      }
    },
...
  }
}